scons optimizations: profiling showed os.walk() and os.path.walk() to be taking a long time...
When looking for tool_ files, walk the source directory only once and stash
the list of tool_*.py files rather than walking every time a tool is loaded.

Move CheckMissingHeaders into the doxygen tool, so a walk() looking for
header files is only run when building apidocs.  No need to look for
missing headers as part of a regular build.

In both uses of walk(), prune .svn directories, since those in turn contain
many subdirectories.  That trims the search space by 80% in aeros.


git-svn-id: http://svn.eol.ucar.edu/svn/eol/common/trunk/site_scons@6223 da89c675-a4d7-0310-8a15-9e7546503515