Add localutils module from nidas to site_scons, where it will automatically be on the python path.
Add armbecross and armcross tools from nidas, except use Detect and WhereIs
instead of Execute, so that scons does not report an error message when
it's only checking to see if the cross-compiler is on the path.

Add kmake.py from nidas, except turn it into a real tool.  Loading the
'kmake' tool now adds the Kmake() builder method to the environment, rather
than replicating that in armbecross and armcross and x86tools.  That makes
x86tools from nidas obsolete.


git-svn-id: http://svn.eol.ucar.edu/svn/eol/common/trunk/site_scons@7005 da89c675-a4d7-0310-8a15-9e7546503515