Add tweaks to svninfo.py so that it works in a directory without any svn files, such as an exported tree.
If there is no svn information in the tree and
the target exists, the target is left alone.

If the target doesn't exist, there are two alternatives:
1. Create a header file, with all the SVN* macros defined as "unknown".
2. Create an empty header file, leaving the macros undefined.

My preference is the second. The developer can then decide
if they want the compile to fail, or can define their own
fallback definitions:
    #include <svnInfo.h>
    #ifndef SVNREVISION
    #define SVNREVISION "final, perfect, absolutely bug-free version!"
    #endif

As a bit of a compromise, it creates a file with the #defines
commented out.



git-svn-id: http://svn.eol.ucar.edu/svn/eol/common/trunk/site_scons@8072 da89c675-a4d7-0310-8a15-9e7546503515