I think the Configure test was trying to build the library dependencies in the LIBS variable, which is what one would expect scons to do, except the check for Qt or QWT does not really depend on those extra libraries. Instead, create the Configure context on a cloned environment and reset LIBS.  This is exactly what we were doing in the netcdf.py tool and others, but I guess I forgot about that.  This fixes the unexpected builds of source tree libraries by the configure checks.  The QWT check is a little kludged in that it must retain the Qt libraries while removing all the others.
Note that it is not enough to create the Configure context on a brand new
clean default Environment (without all the source tree libraries already
added), since that Environment will not have all the right cross-compiler
settings that might be passed down into the tool.  Thus the need for a
cloned Environment with a filtered LIBS.  Unfortunately, that also means
the Configure results cannot be cached by the tool, at least not without
some means to identify which Environments are compiling for the same
target.

I have only tested this in the aeros tree, and not extensively, so there
may still be problems to flesh out.


git-svn-id: http://svn.eol.ucar.edu/svn/eol/common/trunk/site_scons@8461 da89c675-a4d7-0310-8a15-9e7546503515