Fix some issues in the doxygen tool, and cross my fingers I don't break anything for anyone.  None of the uses of the doxygen tool should need to change...
As of doxygen 1.8, the doxytag utility is obsolete and no longer released.
Therefore all the mechanisms for trying to run doxytag on external HTML
documentation have been removed.  If a tag file is not found it is just
ignored.

The dependencies were never right for Doxyfile and Doxygen.  The generated
Doxyfile does not actually depend on the source files, whereas the doxygen
output does.  So now Doxyfile emitter changes the source dependency to the
Doxyfile contents, while the Doxygen emitter gets the source dependencies
by parsing the INPUT parameter in the Doxyfile.  This should be much more
accurate and reliable, except there is one bug somewhere which causes the
Doxyfiles in the aeros tree to be regenerated every time, even though
nothing changed, while everything works correctly in the vardb tree.  At
least now the doxygen builder is more correct in that it does not run
unless the source files or the Doxyfile change.

The LATEX output is disabled by default, since I think HTML is by far the
more commonly used output.

The Doxygen() call is now a pseudo-method which calls the actual
DoxygenBuilder, so the Doxygen() method can add a Clean() target for the
output directory.

All the "if ddebug(): print()" constructs have been replaced with dprint().


git-svn-id: http://svn.eol.ucar.edu/svn/eol/common/trunk/site_scons@9149 da89c675-a4d7-0310-8a15-9e7546503515