Valgrind() pseudo-builder to run valgrind on Command actions
The Valgrind() pseudo-builder accepts scons actions like a regular Command
builder, but it creates two builders, one which includes VALGRIND_COMMAND
and one which sets it to an empty string.  So the same command does not
need to be repeated to run it inside and outside of valgrind, and the
ValgrindLog() analyzer runs automatically on the valgrind output.