several fixes for Valgrind() and ValgrindLog()
The suppressions file has to be passed in the suppressions option with an
absolute path, since we don't know where valgrind will be running.  Allow
extra leader text in the valgrind log file, and try to parse all the
sources as log files until one of them is identified as a valgrind log
file.  Fix the logic in deciding when to run valgrind by default and when
it should be forced by the 'valgrind' global variable.