Improve usage and api of parseconfig module.
The output of pkg-config and other config scripts depends on settings in
ENV like PATH and PKG_CONFIG_PATH, so the config output is now only cached
with each Environment instance and not in the global VariableCache.

Add a ParseConfig() function to the parseconfig module as an alternative to
Environment.ParseConfig().  If the config script succeeds, then the flags
are merged into the environment and the function returns true.  Tools can
use this to test a config script and use the settings if it works, and
otherwise continue with other config methods.

The CheckConfig() return value has been inverted. Rather than return the
actual return the actual return code, just return True if the return code
is zero, meaning the config script settings exist.

Try to use ParseConfig() and CheckConfig() more consistently in place of
direct calls to subprocess.Popen().

Add 'env' as a parameter to QwtTool.findPkgConfig().  The last change added
env[ENV] to the pkg-config call, but there was no 'env' variable.