Run the QtCore Configure check only once per Environment.
The Qt4 modules and tools may require QtCore multiple times in the same
environment, causing the Configure() check for QtCore to run multiple times
even though they all return the same answer, and all the checks print out
the message 'Checking for C++ library QtCore...'  Now the first result is
cached in the Environment.  This could be further sped up by caching the
result across all Environments, but that seems not worth the risk at this
point.