convert env['ENV'] values to strings before passing to Popen
execve() raises an exception if any of the keys or values passed in the
environment dictionary are not strings, so scons must convert them first.
This change now does the same thing to the environment passed into Popen.
Recall that we want to use the process environment which scons provides for
calls to pkg-config and other config scripts, for consistency with
variables like PKG_CONFIG_PATH.