fix doxygen tagfile dependencies and other document integration
Add tagfiles as implicit sources and targets.  Since has been impossible
for a while now to generate tagfiles from existing html source, all
tagfiles with explicit references must exist or otherwise are ignored, but
dependencies for generated tagfiles make sure that dependent subproject
apis are built before the parent project.  The new convention is that
references are added to DOXREF as construction variables so the project can
specify them if desired, otherwise they are ignored.