remove import of builtins and use of 'future' module
RHEL 7.4 is python 2.7.5, which apparently does not have the builtins
module yet.  However, scons 2.3.0 has a workaround for that, and so scons
code which imports builtins would still work with python 2.7.5.  On
machines which update scons to a newer version without the builtins
compatibility, say 2.5.1, the builtins import no longer works.  I don't
think the code really needs to explicit import of str or object from
builtins, they do the right thing in either 2.7.5 or 3.x without using
builtins, so this commit tests that theory.