refactor qt5 to fix conflation of module names and packages
Qt module names do not contain the version number, so that is enforced now
in the use of EnableQtModules(), and tool names correspond exactly to the
module name without the version number.  The module name is now used more
consistently, and differentiated more clearly from the pkg-config package
names and library files which *do* contain the version.  Specifically:

 * Remove tool names with qt5 and use unqualified tools everywhere.
 * Error if a module name passed to EnableQtModules() contains a version.
 * Keep the qualified name to find the pkgconfig package.
 * Keep the qualified library name when not using pkgconfig.
 * When not using pkg-config, add the cpp definition QT_<MODULE>_LIB
   as provided by the pkg-config packages.
 * The qt module tools do not try to infer the Qt version, instead they
   just pass the corresponding module and its dependencies to
   EnableQtModules().  The Qt version must be setup elsewhere.
 * Refactor the qt5 EnableQtModules() itself into different methods to
   enable a Qt module on each platform: osx, linux, windows.