implicitly load qt4 or qt5 if QT_VERSION set
QT_VERSION must be set before requiring a Qt module tool, either by first
requiring the qt4 or qt5 tool, or by setting it in the environment with a
project or global tool.  Setting QT_VERSION then implicitly requres the
right qt4 or qt5 tool, or else raise an error if it is not set yet.