fix boost_python tool on Fedora 31
The naming of the boost_python library varies on systems where it can be
built against either python2 or python3, so this tool now attempts to
figure out the "default" python-config script to use and derive the right
boost_python library from that.  This has not been tested on RHEL yet.

This also moves the ldd() function into a module where it can be shared by
both the boost_python and deploy tools.