no point trying to build against source when the location is unknown
Since the path to the raf source inside a source tree is not known relative
to the environment to which the tool is being applied, there is no point in
trying to add the right include path.  Likewise for the library.  This
reverts this tool to working only for the installed raf, by adding raf to
the LIBS.  The in-source raf tool will be found using a tool_raf.py.