make sure OPT_PREFIX paths show up last in CPPPATH and LIBPATH
OPT_PREFIX is usually used as a fallback to find dependencies which are not
in the source tree or not in the usual system locations, so now it only
appears at the end of the includes and library paths.  This may break
builds which expect OPT_PREFIX to take precedence over all other
dependencies, but I'm thinking that needs to be handled a different way.
When a source tree needs to build against an OPT_PREFIX to which the source
has been installed before, then the build could use the wrong headers and
libraries if OPT_PREFIX precedes the in-source paths.  This is especially a
problem for headers and dynamic libraries, since static libraries at least
can be added to LIBS with the full file path.

If there are other cases where fallback paths need to be added to LIBPATH
and CPPPATH, then it would be reasonable to add AppendCppLastPath() and
AppendLibLastPath() as Environment methods.