move postgres_testdb classes to a module in eol_scons
There is an apparent bug in python 3.6 which causes the warnings.py tool to
be loaded as a module by the subprocess package when the postgres_testdb.py
script was run from the tools directory.  This is an attempt to avoid
conflict by moving all the basic functionality into the eol_scons package
and out of the tools directory, and it just generally seems like a good
idea to make the functions accessible without the tool.

This includes some pylint cleanup.  Builds which call the
postgres_testdb.py script directly with a hardcoded path will have to be
updated.  Builds which use the POSTGRES_TESTDB_PATH variable should still
work.