detect and warn about cyclic tool loads
If a tool_ file creates an Environment which requires the
same tool function before the tool function has been
exported, then an infinite recursion happens, with eol_scons
trying to load the file containing the tool function
over and over.  Instead, detect this situation and
print a warning about it.