set python-config path with a variable
For systems with both python3-config and python-config on the path,
add PYTHON_CONFIG variable to pick one.  Also, if python-config
does not report a python shared library to link to, or if it does,
no libboost_python can be found which depends on that library,
then just take whatever is found first.  Anaconda installs may
not report a python library in python-config --libs, and the
boost_python library may not list a python library as a dynamic
dependency.  So really that leaves not much choice.  And even
though a compile may seem to work against anaconda3 python,
loading the module still fails with missing symbol errors, and that
is still a mystery.