fix for scons 4.4 by removing warnings tool
scons 4.4 uses an os module from newer python (eg 3.10) which
imports a module named warnings, and that collides with the
warnings tool.  since the warnings tool and also the debug, optimize,
and profile tools are just hooks for buildmode, remove them and
call the hooks directly in the buildmode tool.  probly i thought
implementing compiler flags as tools would allow them to be
overridden or expanded, but since that has never been done in practice,
afaik, this approach is much simpler and avoids the collision
with the warnings module.