refactor version handling
defer the derivation of shared library major and minor versions to the
emitter, in case they are not yet set in the generate() function when
tool is loaded.  rather than try to extract the library name by
matching and replacing the version suffix, pass the unversioned name
to the emitter and build the versioned names by appending the versions.
make installation work with --install-sandbox by creating the symlinks
relative to the actual installed library node, rather than the target
passed into the install.  be more flexible about parsing major and minor
versions from REPO_TAG.

it looks like the latest shared library builders in SCons do everything
this tool does except for setting ARCHLIBDIR, in which case projects
should start phasing this tool out.