add EnableInstallAlias(), fix for --install-sandbox
allow disabling eol_scons override of Install(), for projects which
want to use different aliases for different installs, such as
install, install.root, install.doc.  try to support --install-sandbox
better by passing the destination directory as a string and not a
node, since nodes are not regenerated with the sandbox prefix.