modify Configure context, only check for qt5 once
it seems to sometimes not be enough to replace LIBS in a clone
and then create the Configure context, since it is not certain the
context environment will be a copy.  so replace LIBS in the
context environment explicitly.

also, only check for qt5 once, on the presumtion that all other
checks will return the same.