resort to git clean for source copies
when trying to generate a tar file from a copy rather than a clone, there is
just too much extra stuff that can be included unnecessarily, so resort to a
full git clean in the copy.  this means all source files have to be committed
to git to be included in the tar file, but they can still have pending
changes.