incremental documentation updates
fix doxygen configuration to generate doc entries for tool modules, and also
to process doxygen formatting in python docstrings.  move the doxygen tool
documentation into the tool and reference it from the README.  add docs alias
to run doxygen.  The CHANGELOG is included as a doxygen input, even though
doxygen does not seem to format reference links in headers correctly.