simplify and clarify SpawnerLogger
Make it callable so the instance itself can be assigned to the SCons SPAWN
environment variable, and allow it to be used more easily without LogAction by
implicitly opening the log file if not open already.  Make the patterns
argument more consistent: None means use the default, empty list [] means
nothing will match and so nothing will be written, and [r'.*'] must be passed
to allow all output.  Also, stdout will be filtered even if no log path is
specified, since it makes sense that the two can be independent.  This may
change prior behavior for the few, if any, places where a LogAction was
created without a log path.  However, it allows a LogAction to be used to
filter output even if a log file is not desired.