fix container deploy problem for qt5 libxqcb.so
at least in a podman container, the SCons Copy() action seems to want to do
something when copying the qt5 libxqcb.so which fails with an error, maybe
trying to call stat() on it or something (this sounds vaguely familiar...).
anyway, to get around the failure for deploy targets, just use the cp command
for the deploy target same as for all the shared libraries.