find netcdf-cxx4 installed with cmake on fedora
since the netcdf-cxx4 library when built with cmake from source ends up with a
different name, go ahead and search for both names when the NETCDFCXX4DIR
variable is set, and in both subdirectory names, lib and lib64.  as before,
use the default library name for the platform unless pkg-config works.