reduce the number of pkg-config calls
cache the qt5 header dir after finding it once, rather than
calling pkg-config for each qt module which needs it.  this
presumes all the scons environments will be building against
the same Qt5, which seems like a safe presumption.

the regular expression search for Q_OBJECT is slow.  the only
mitigation so far is to compile the RE only once.

it seems unnecessary to first check for a qt module in pkg-config
and then call pkg-config again to get the flags, so just
do it once and expect some output to know if it worked.