shortcut SCons AddMethod to avoid wrappers
the SCons.Util.AddMethod implementation handles lots of cases and can end
up wrapping a function to make it a method rather than just binding it
directly to the instance.  besides seeming to slow down performance, it
makes profile call stacks more difficult to trace.  so try to optimize
method binding by 1) not requiring the function to be redefined with a
different name before being added, and 2) binding directly to the instance
without calling SCons.Util.AddMethod to avoid any chance of a wrapper.

I also played with binding directly to the class of the passed environment.
While I could get it to work, it did not seem to change performance, and
that would introduce the complication of deciding which methods were
"tool-specific" and belonged to an instance, and which were really
eol_scons extensions that should apply to every environment instance.