Add default include/ and lib/ to PATH
This covers the case of ARM64:

Add brewPrefix/include to CPPPATH and brewPrefix/lib to LIBPATH.
There are a number of packeges that don't have a pkg-congig .pc file
just for Mac/homebrew (e.g. netcdf-cxx4 and xerces-c).

This will cover the default case in other unix environments of
/usr/local/include and /usr/local/lib

Alternative is to go into each tool and add code there.